fix: Correct setter method names and parameter types in Accommodation and Flight classes

+++ backend/src/main/java/com/roamroute/backend/entity/Accommodation.java
@@ -82,7 +82,7 @@ public class Accommodation {
 		return latitude;
 	}
 
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
@@ -90,7 +90,7 @@ public class Accommodation {
 		return longitude;
 	}
 
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 }
